+++ api/shared/agent/tools/index.js
@@ -3,6 +3,9 @@ const createGetNextBirthdayTool = require("./getNextBirthdayTool");
 const createGetContactsTool = require("./getContactsTool");
 const createGetDayConfigurationsTool = require("./getDayConfigurationsTool");
 const createAddDayConfigurationsTool = require("./addDayConfigurationsTool");
+const createGetEventsTool = require("./getEventsTool");
+const createAddEventTool = require("./addEventTool");
+const createUpdateEventDetailsTool = require("./updateEventDetailsTool");
 
 module.exports = function createAgentTools(context) {
   const toolModules = [
@@ -11,6 +14,9 @@ module.exports = function createAgentTools(context) {
     createGetContactsTool(context),
     createGetDayConfigurationsTool(context),
     createAddDayConfigurationsTool(context),
+    createGetEventsTool(context),
+    createAddEventTool(context),
+    createUpdateEventDetailsTool(context),
   ];
   const definitions = toolModules.map((tool) => tool.definition);